requirements.txt
mlflow==2.14.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
PyYAML==6.0.1
pytest==8.3.3
